requirements.txt
# Requirements for the project
flask
flask-cors
openai
python-dotenv
